refactor(OmokListener): 리스너 네이밍, 사용하지 않는 클래스 삭제
